坐标类型改为float

+++ MO.Servers/MO.GrainInterfaces/Game/IRoom.cs
@@ -13,6 +13,6 @@ namespace MO.GrainInterfaces.Game
         Task PlayerEnterRoom(IUser user);
         Task PlayerLeaveRoom(IUser user);
         Task PlayerReady(IUser user);
-        Task PlayerGo(IUser user, double x, double y);
+        Task PlayerGo(IUser user, float x, float y);
     }
 }
